Added Sacrifice, corrected Prana View and added Prana UI View (for GUI)

+++ Assets/Scripts/Common/Constants.cs
@@ -4,7 +4,7 @@ namespace Core
     {
         public const int CitiesLayer = 6;
 
-        public const int VirtueValueMax = 100;
+        public const float VirtueValueMax = 100f;
         public const int PranaValueMax = 100;
     }
 }
